API keys: move flag-aware capped verdict out of KeyResource (fix import cycle)
Importing getFeatureFlags into KeyResource created a lib/auth → key_resource →
lib/auth cycle (biome noImportCycles, failing format-check). Move the flag-aware
per-key capped computation to getApiKeysSpendCappedByModelId in
lib/api/keys/spend_limit; the keys list endpoint computes the map and passes it
to KeyResource.toJSONWithSpaces. toJSON falls back to the persisted credit state
(flag-off behavior) for the single-key mutation responses.
